Split cohort api user management files to reduce large amounts of code in one file
