+++ qa-portal-angular/package.json
@@ -13,21 +13,21 @@
   "dependencies": {
     "@angular/animations": "8.0.2",
     "@angular/cdk": "8.0.1",
-    "@angular/common": "~8.0.1",
-    "@angular/compiler": "~8.0.1",
-    "@angular/core": "~8.0.1",
-    "@angular/forms": "~8.0.1",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "^8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/forms": "^8.0.3",
     "@angular/http": "^7.2.15",
     "@angular/material": "8.0.1",
-    "@angular/material-moment-adapter": "^8.2.0",
-    "@angular/platform-browser": "~8.0.1",
-    "@angular/platform-browser-dynamic": "~8.0.1",
-    "@angular/router": "~8.0.1",
-    "@types/jest": "^24.0.18",
-    "angular-calendar": "^0.27.18",
+    "@angular/material-moment-adapter": "^8.2.3",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "^8.0.3",
+    "@angular/router": "^8.0.3",
+    "@types/jest": "^24.0.23",
+    "angular-calendar": "^0.27.21",
     "angular-in-memory-web-api": "^0.9.0",
-    "bootstrap": "^4.3.1",
-    "chart.js": "^2.8.0",
+    "bootstrap": "^4.4.1",
+    "chart.js": "^2.9.3",
     "date-fns": "^1.29.0",
     "hammerjs": "^2.0.8",
     "http-status-codes": "1.3.2",
@@ -39,29 +39,29 @@
     "ngx-toastr": "10.0.4",
     "popper": "^1.0.1",
     "randomcolor": "^0.5.4",
-    "rxjs": "^6.4.0",
-    "simplebar": "^4.1.0",
+    "rxjs": "^6.5.3",
+    "simplebar": "^4.2.3",
     "tslib": "^1.9.0",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular-devkit/build-ng-packagr": "~0.800.0",
-    "@angular/cli": "~8.0.3",
-    "@angular/compiler-cli": "~8.0.1",
-    "@angular/language-service": "~8.0.1",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular-devkit/build-ng-packagr": "^0.800.6",
+    "@angular/cli": "^8.0.6",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/language-service": "^8.0.3",
+    "@types/jasmine": "^3.3.16",
+    "@types/jasminewd2": "^2.0.8",
     "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^5.2.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.1.0",
+    "ng-packagr": "^5.7.1",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tsickle": "^0.35.0",
