+++ qa-portal-services/cohort-api/src/main/java/com/qa/portal/cohort/services/CohortManagementService.java
@@ -1,7 +1,7 @@
 package com.qa.portal.cohort.services;
 
 import com.qa.portal.cohort.keycloak.KeycloakCohortResourceManager;
-import com.qa.portal.cohort.keycloak.KeycloakUserResourceManager;
+import com.qa.portal.cohort.keycloak.KeycloakUserCohortManager;
 import com.qa.portal.common.dto.QaCohortDto;
 import org.springframework.stereotype.Service;
 
@@ -9,33 +9,34 @@ import org.springframework.stereotype.Service;
 public class CohortManagementService {
 
     private KeycloakCohortResourceManager keycloakCohortResourceManager;
-
-    private KeycloakUserResourceManager keycloakUserResourceManager;
+    
+    private KeycloakUserCohortManager keycloakUserCohortManager;
 
     private CreateCohortOperation createCohortOperation;
 
     private UpdateCohortOperation updateCohortOperation;
 
     public CohortManagementService(KeycloakCohortResourceManager keycloakCohortResourceManager,
-                                   KeycloakUserResourceManager keycloakUserResourceManager,
                                    CreateCohortOperation createCohortOperation,
-                                   UpdateCohortOperation updateCohortOperation) {
+                                   UpdateCohortOperation updateCohortOperation,
+                                   KeycloakUserCohortManager keycloakUserCohortManager) {
         this.keycloakCohortResourceManager = keycloakCohortResourceManager;
-        this.keycloakUserResourceManager = keycloakUserResourceManager;
+        this.keycloakUserCohortManager = keycloakUserCohortManager;
         this.createCohortOperation = createCohortOperation;
         this.updateCohortOperation = updateCohortOperation;
     }
 
     public QaCohortDto createCohort(QaCohortDto cohortDetails) {
+    	System.out.println(cohortDetails);
         createCohortOperation.createCohort(cohortDetails);
         keycloakCohortResourceManager.createCohort(cohortDetails.getName());
-        keycloakUserResourceManager.updateCohortMembers(cohortDetails);
+        keycloakUserCohortManager.updateCohortMembers(cohortDetails);
         return cohortDetails;
     }
 
     public QaCohortDto updateCohort(QaCohortDto cohortDetails) {
         updateCohortOperation.updateCohort(cohortDetails);
-        keycloakUserResourceManager.updateCohortMembers(cohortDetails);
+        keycloakUserCohortManager.updateCohortMembers(cohortDetails);
         return cohortDetails;
     }
 }
